+++ docs/changelog.rst
@@ -155,7 +155,9 @@ Backend API
 Models
 ------
 
-- (no changes yet)
+- Remove ``.copy()`` method on all model classes.
+  Use the ``.replace()`` method instead.
+  (Fixes: :issue:`1464`, PR: :issue:`1774`)
 
 Extension support
 -----------------